+++ src/test/kotlin/racingcars/FixedRandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcars
 
 class FixedRandomNumberGenerator(private val number: Int) : RandomNumberGenerator {
     override fun generate(): Int = this.number
